% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttp.fun.R
\name{af.fun}
\alias{af.fun}
\title{Calculates Asymmetry Factor of peaks}
\usage{
af.fun(z)
}
\arguments{
\item{z}{A data.frame containing all peaks as columns and timepoints in column 1 from a csv file}
}
\value{
The AF value for the peak
}
\description{
Calculates AF of peaks
}
\examples{
add_numbers(1, 2) ## returns 3
}
