% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.fun.R
\name{area.fun}
\alias{area.fun}
\title{Calculates AUC of peaks}
\usage{
area.fun(z)
}
\arguments{
\item{z}{A data.frame containing all peaks as columns and timepoints in column 1 from a csv file}
}
\value{
The Area under the Curve
}
\description{
Calculates AUC of peaks
}
\examples{
add_numbers(1, 2) ## returns 3

}
