% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwhm.fun.R
\name{fwhm.fun}
\alias{fwhm.fun}
\title{Calculates FWHM of peaks}
\usage{
fwhm.fun(z)
}
\arguments{
\item{z}{A data.frame containing all peaks as columns and timepoints in column A from a csv file}
}
\value{
The FWHM value for the peak
}
\description{
Calculates FWHM of peaks
}
\examples{
add_numbers(1, 2) ## returns 3

}
