% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrt.fun.R
\name{hrt.fun}
\alias{hrt.fun}
\title{Calculates HRT of peaks}
\usage{
hrt.fun(z)
}
\arguments{
\item{z}{A data.frame containing all peaks as columns and timepoints in column 1 from a csv file}
}
\value{
The HRT value for the peak
}
\description{
Calculates HRT of peaks
}
\examples{
add_numbers(1, 2) ## returns 3
}
