% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localMaxima.fun.R
\name{localMaxima.fun}
\alias{localMaxima.fun}
\title{Finds local maxima on peak traces}
\usage{
localMaxima.fun(z)
}
\arguments{
\item{z}{}
}
\value{

}
\description{
Finds local maxima on peak traces
}
\examples{
test <- localMaxima(z)

}
