% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localMinima.fun.R
\name{localMinima.fun}
\alias{localMinima.fun}
\title{Finds local minima on peak traces}
\usage{
localMinima.fun(z)
}
\arguments{
\item{z}{}
}
\value{

}
\description{
Finds local minima on peak traces
}
\examples{
test <- localMinima(z)

}
