% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakDim.fun.R
\name{peakDim.fun}
\alias{peakDim.fun}
\title{Calculates Peak Characteristics}
\usage{
peakDim.fun(z)
}
\arguments{
\item{z}{A data.frame containing peak values in column 2 and timepoints in column 1 from a csv file}
}
\value{
The FWHM, HRT and AUC value for the peaks
}
\description{
Calculates Peak Characteristics
}
\examples{
test <- fwhm(one)
names(test) <- c("xmax", "ymax", "ybase", "halfMax", "xALow", "yALow", "xAHigh", "yAHigh", "xBHigh", "yBHigh", "xBLow", "yBLow", "mA", "mB", "aA", "aB", "xA", "xB", "fwhm", "hrt", "auc")
}
