% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttp.fun.R
\name{tf.fun}
\alias{tf.fun}
\title{Calculates Tailing Factor of peaks}
\usage{
tf.fun(z)
}
\arguments{
\item{z}{A data.frame containing all peaks as columns and timepoints in column 1 from a csv file}
}
\value{
The TF value for the peak
}
\description{
Calculates TF of peaks
}
\examples{
add_numbers(1, 2) ## returns 3
}
