% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttp.fun.R
\name{ttp.fun}
\alias{ttp.fun}
\title{Calculates TTP of peaks}
\usage{
ttp.fun(z)
}
\arguments{
\item{z}{A data.frame containing all peaks as columns and timepoints in column 1 from a csv file}
}
\value{
The TTP value for the peak
}
\description{
Calculates TTP of peaks
}
\examples{
add_numbers(1, 2) ## returns 3
}
